pyproject.toml
[tool.poetry]
name = "compras-do-mes"
version = "0.1.0"
description = "App back-end em RestFul de lista de compras"
authors = ["Edimar Freitas de Sá <edimarfreitas95@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.3"
djangorestframework = "^3.15.1"
Markdown = "^3.6"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
uuid = "^1.30"
psycopg2 = "^2.9.9"
djoser = "^2.2.2"
bcrypt = "^4.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pip = "^24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
